src: improve log message when user is not logged in
It's debatable whether this is a great improvement. I will adjust it
if it turns out to be not ideal. Now, when a user is not logged in,
the CLI writes the caught exception message and status code to
the output log. It would be ideal if we could determinen whether the
user was already logged in before making a remote call, however, so
that we can make the output nicer looking. In any case, here is what
the error looks like now.

```sh
~/s/nodejs-rest-http (master $=) nodeshift --osc.strictSSL=false .
2017-11-16T16:58:53.841Z INFO loading configuration
2017-11-16T16:58:53.867Z INFO using namespace myproject at https://192.168.99.100:8443
2017-11-16T16:58:53.871Z INFO enriching Deployment resource
2017-11-16T16:58:53.873Z INFO enriching Route resource
2017-11-16T16:58:53.874Z INFO enriching Service resource
2017-11-16T16:58:53.879Z INFO openshift.yaml and openshift.json written to /Users/lanceball/src/nodejs-rest-http/tmp/nodeshift/resource/
2017-11-16T16:58:53.883Z INFO creating archive of package.json, app.js, public, bin, LICENSE
2017-11-16T16:58:53.922Z ERROR User "system:anonymous" cannot get buildconfigs in project "myproject"
2017-11-16T16:58:53.922Z ERROR Status code 403
```

I also added debuglog for the CLI, so if the command is run as

```sh
$ NODE_DEBUG=nodeshift.cli nodeshift .
```

then it will print the stack and config object to the console.

Fixes: https://github.com/bucharest-gold/nodeshift/issues/98