update(enrichers): refactor of how the resource files are enriched.
Instead of looping on the file list, we now loop on a list of enrichers.  Which will map and enrich the resources.

This led to a lot less duplicated code in each of the resource enricher files.

This commit also includes the update for creating a missing Service resource file and a missing Deployment resource file