feat: add the --deploy.env flag
This will allow users to pass in environment variables that should be added to a deployment config.

ex: nodeshift --deploy.env NODE_ENV=development

If a deployment.yaml in the .nodeshift directory already exists with environment variables, they will be merged together
If there are duplicate vars, the ones passed in from the cli/api will take precedence.

fixes #223