feat: remove deprecation warnings from openshift rest client
BREAKING CHANGE: The api for the openshift rest client has changed slightly, but there should be no nodeshift api changes

While this doesn't have to be a semver-major release, there might be some unexpected bugs.  One known issue, is we are no longer passing a custom config to the rest client.  If this removal causes issues we can work on a way to put it back in

fixes #377